moved comands into subfolders
